Organize files, fix warnings and typos
